Adds plastitanium shards, and slightly modifies plasma shards (#3600)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Credit: Naksu, octareenroon91, and vuonojenmustaturska for the original
plasma shard sprites.
Original PR is here:
[https://github.com/tgstation/tgstation/pull/40327](https://github.com/tgstation/tgstation/pull/40327)

This change adds plastitanium shards and lightly modifies plasma shards.

**Plastitanium shards**
| plastitaniumlarge | plastitaniummedium | plastitaniumsmall |
plastitaniumtiny |

|-------------------|--------------------|-------------------|------------------|

|![image](https://github.com/user-attachments/assets/2196d5be-b8f5-4879-bc00-20b92a25de5c)|![image](https://github.com/user-attachments/assets/341a3a00-1213-4d7e-9792-5ac47ba70539)|![image](https://github.com/user-attachments/assets/91ac24fb-b9bd-415d-804c-07bbb3a54546)|![image](https://github.com/user-attachments/assets/9424b468-fbd9-4507-8db3-94becfe45272)|

**Plasma shards**
| before
|![image](https://github.com/user-attachments/assets/553d1c11-9ecc-401e-869c-0cd9701ca015)|
|--------|---|
| after
|![image](https://github.com/user-attachments/assets/e4f02d01-628d-443c-8765-4e457ab1a13c)|


## Why It's Good For The Game

Plasma shards were modified because it seemed to me like there were a
few artifacts that remained in the sprite. These have been cleaned up.

As for the plastitanium shards, they have been added in order to make
things a little more consistent. Currently, if a plastitanium window is
broken, plasma shards are dropped. This change should fix that and make
it so that if a plastitanium window is broken, a _plastitanium_ shard is
dropped.

![image](https://github.com/user-attachments/assets/512dc9a9-9b17-4a49-bece-dd583bacd977)

All of the glass shard sprites will likely need to be touched up in the
near future, but this may serve as an okayish backbone for future work.
My spritework here may not be the best, so please critique it if it can
be improved.

## Changelog

:cl:
add: Added plastitanium shards
fix: fixed some artifacts in the plasma glass shard sprites
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->