Adds rail lights to inteq ships. (#3605)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Title. Also tweaks the cargo order - it costs 100 instead of 250 now.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
- I think rail lights have a great vibe for inteq and would fit them
well. Probably fitting for other factions too but I'm lazy.
- Price reduced because, ima be real, 250 for a funny light, while More
Important Gear costs the same, is NOT worth it.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: All the inteq ships (and the beluga inteq office) now start with a
rail light for every gun.
tweak: Rail light crate now costs 100 credits.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->