PGF - Elated Bolide Class Assault Lander (#3399)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds the Elated Bolide Class and assorted fluff that I made for it. 

![StrongDMM-2024-09-28 22 47
07](https://github.com/user-attachments/assets/8b2eeff5-32c4-40df-ac90-fae97314c278)
![StrongDMM-2024-09-28 22 47
08](https://github.com/user-attachments/assets/16354732-bb0a-47ef-9e14-2b70a640a72e)


The Elated Bolide is named for the excitement that it's crew experiences
when hitting a hostile atmosphere, and the way that a particularly hard
entry can appear from the ground.
In the frontier, the Bolide-Class is used both in exploration and
civilization building (imperialism) with the marines being equipped with
a surplus of equipment to quickly break down and salvage unwanted
structures (enemy fortifications) while building colonial
infrastructure, such as concrete roads (bunker complexes).
6 Marines are carried, 4 of which are marine engineers. The traditional
assignment is splitting the marines into 2 squads of 3, although in the
end it is up to the Captain and Marine LT to decide how the teams are
organized. Several fluff papers have been included to try and encourage
roleplay aboard the vessel, as well as a degree of paperpushing.

The Bolide is well defended from the sides thanks to it's Etherbor Point
Defense Systems, with the rear mounts carrying a lighter deterrent
turret.
The Bolide is also well armed thanks to it's armory, able to stock all 6
marines (assuming that some marines are willing to carry Tools of some
flavor instead of a large gun), and contains some specialist gear for
sensitive explosives defusal. And explosives planting.

TLDR:
- High powered turrets near the cargo bay
- Surplus of basic engineering equipment
- Decently sized armory
- Decently sized cargo bay
- Lack in medical, advanced engineering equipment.
- Lots of demolition tools. 
<details><summary>Pictures</summary>
<p>


![dreamseeker_sJpVpK4vye](https://github.com/user-attachments/assets/8a497c4c-5de0-474a-8029-c4c96640c395)

![dreamseeker_pAFuv4kWM6](https://github.com/user-attachments/assets/ebc214aa-6296-4e8c-831f-39d4a430e1bf)

![dreamseeker_KjBGWSFeXV](https://github.com/user-attachments/assets/40da87fd-3dd3-4a4d-913b-9142f11bde0e)

![dreamseeker_hVrb3EOixZ](https://github.com/user-attachments/assets/dd17472f-295f-4dd5-9b21-65ecdb1dc28b)

</p>
</details> 



<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
's PGF content. Needs a bit more decor though but I'm at a loss for how
right now. Technically completely playable.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: A new PGFN vessel has entered the frontier - the Elated Bolide
class makes a bright entry into the local system.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Erika Fox <94164348+Bjarl@users.noreply.github.com>
Co-authored-by: meem <75212565+meemofcourse@users.noreply.github.com>