PGF/Crying Sun-housekeeping again (#3625)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds/fixes a number of things regarding the Crying Sun in particular and
makes a small change to the Gezena outfits.

To be precise:

The Marine Lt. role is now officially considered an officer.

The Helmsman role's ID now actually displays "Helmsman"

The Marine Lt. and the Captain both had their lockers re-organized and
had their loose etherbor pistols swapped with pistols responsibly stored
in proper cases.

The Captain now has the megaphone that was supposed to be in their
quarters oh-so long ago.

The Crying Sun’s refrigerator now has some water bottles in it.
Hydration is important. Removed the beer. Marines and Sailors on active
duty shouldn't be drinking. Just having it on board is inviting trouble.
If they want to sneak in a drink, let them smuggle it on board
themselves. Makes for fun RP.

The Crying Sun's lounge no longer has a cigarette machine as smoking
should only be done in designated smoking areas while on a Navy ship.
Being this is a spaceship, we agreed that the designated space is
**off** the ship.

![Screenshot 2024-10-22
182426](https://github.com/user-attachments/assets/53fccb93-f601-4784-a5dc-627e3f4d8266)
![Screenshot 2024-10-23
013125](https://github.com/user-attachments/assets/2e9786bf-a8a5-47a1-b45f-fcf83128c2c7)
![Screenshot 2024-10-23
013106](https://github.com/user-attachments/assets/76d15a3e-019b-45d4-a051-e6434fd5684f)
![Screenshot 2024-10-23
123204](https://github.com/user-attachments/assets/cda4d290-95e8-40eb-b753-400bd2c823cd)
![Screenshot 2024-10-23
123219](https://github.com/user-attachments/assets/8ad7e791-883a-4418-acbe-c20aa4422977)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
All of these changes were discussed in the ruinyard prior to their
inclusion here barring the on board drink changes.

Tidying up storage and keeping ship code / design congruent with new
standards keeps ships up to date and prevents the need for large, labor
intensive update sweeps.

Removal of the Cigarette vendor from the ship fits the vessel's navy
theme and also gives crew a good reason to run down to the outpost and
increase the likelihood of interactions with other vessels.

In regards to the water inclusion and beer removal. This is good for RP
as it brings the ship's supplies in line with real Navy regulation. The
ship already has strong spirits on board under the control of the
medical doctor. Something that has historical merit in the Navy. Loose
and freely accessible alcohol in the mess hall is frankly nonsensical
given the theme of the ship.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Improved miscellaneous supplies on the Crying Sun
add: Added a blackbox recorder and set of handrails to the Crying Sun 
fix: Changed out the loose beampistols with cased versions on Crying Sun
fix: Fixed Crying Sun roles to display properly
del: Removed the beer ring and cigarette machine from the Crying Sun
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: DIB-DOG <DaneOfThieves@gmail.com>